chore: Update kirei/_app/cli.py to use os.listdir for path completion
The code changes in kirei/_app/cli.py modify the completer for the "文件(需要输入文件路径)" parameter in the _user_file_inquirer function. The completer is updated to use os.listdir instead of the default PathCompleter. This change is made to provide a list of paths for path completion based on the files in the current directory.

This commit message follows the established convention of using a prefix to indicate the type of change (chore for chore) and providing a concise summary of the changes made.