Update completer for "文件(需要输入文件路径)" parameter in _user_file_inquirer function
The code changes in kirei/_app/cli.py update the completer for the "文件(需要输入文件路径)" parameter in the _user_file_inquirer function. The completer is modified to use os.listdir instead of the default PathCompleter. This change is made to provide a list of paths for path completion based on the files in the current directory.